Update with few changes for sls and multiaccount
