+++ core/jazz_ui/e2e/page-objects/jazzservices.po.ts
@@ -107,7 +107,7 @@ export class Jazz {
     return element(by.xpath('//div[contains(text(),"successful")]'));
   }
   getAssetStatusVerify() {
-    return element(by.xpath('//div/div[1]/div[2]/div/div[2]/ul[1]/li[1]/div[2]'));
+    return element(by.css('div.asset-content-row.row ul.section-left.col-md-6.col-sm-12 li:nth-child(1) > div.det-value'));
   }
   getLamdaName() {
     return element(by.xpath('//div[@class="table-row pointer"]/div[text()="function"]/preceding-sibling::div'));
@@ -417,7 +417,7 @@ export class Jazz {
     return element(by.xpath('//div[contains(@class,"each-section-accReg")]//button[contains(@class,"btn dropdown-btn dropdown-toggle")]'));  
     }
   regionSelect() { 
-    return element(by.xpath('//div[contains(@class,"col-lg-8 col-md-8")]//div[2]//div[2]//dropdown[1]//div[1]//button[1]//span[2]'));  
+    return element(by.xpath('//*[@id="typeform"]/section[4]/div/div[2]/section/div/div[2]/div[2]/dropdown/div/button'));  
     }
   firstAccount() { 
     return element(by.xpath('//section[4]/div/div[2]/section/div/div[1]/div[2]/dropdown/div/ul/li[1]/a'));  
@@ -430,6 +430,16 @@ export class Jazz {
     }
   westRegion() { 
     return element(by.xpath('//a[contains(text(),"us-west-2")]'));  
+    }
+
+  deployDesciptor() { 
+    return element(by.xpath('//section[@id="deployment-descriptor"]//label'));  
+    }
+  eastRegionVerify() { 
+    return element(by.xpath('//div[contains(text(),"us-east-1")]'));  
+    }
+  westRegionVerify() { 
+    return element(by.xpath('//div[contains(text(),"us-west-2")]'));  
     }
 
 }
